requirements.txt
flask==2.3.3
openai==0.27.8
googletrans
langdetect
python-dotenv
